package-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
requests==2.31.0
python-consul==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0